requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beanie==1.30.0
certifi==2025.7.9
click==8.2.1
Cython==3.1.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lazy-model==0.2.0
motor==3.7.1
passlib==1.7.4
ply==3.11
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyzk==0.9
rsa==4.9.1
setuptools==75.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
thriftpy2
tomli==2.2.1
typing_extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
